feat: type alias for fp

+++ src/code/math/pi_23.hpp
@@ -9,7 +9,7 @@ namespace tifa_libs::math {
 
 inline u64 pi_23(u64 n) {
   if (n < 2) return 0;
-  static auto div = [](u64 n, u64 p) -> u64 { return (u64)((double)n / (double)p); };
+  static auto div = [](u64 n, u64 p) -> u64 { return u64((f64)n / (f64)p); };
   u64 n2 = isqrt(n), n3 = ikth_root(n, 3), n6 = ikth_root(n, 6), n23 = n / n3;
   auto prime = prime_seq((u32)n2 + 1000);
   u32 pidx = 0;
